compat: move `running_on_ci()` to compat
Currently defined in `_pytest.assertion.util`, but then imported from
`_pytest.terminal`, so makes sense to move it to the general utils
place.
