+++ src/_pytest/compat.py
@@ -304,3 +304,9 @@ class CallableBool:
 
     def __call__(self) -> bool:
         return self._value
+
+
+def running_on_ci() -> bool:
+    """Check if we're currently running on a CI system."""
+    env_vars = ["CI", "BUILD_NUMBER"]
+    return any(var in os.environ for var in env_vars)
